requirements.txt
anthropic>=0.40.0
tavily-python>=0.3.0
pydantic>=2.0.0
streamlit>=1.35.0
apscheduler>=3.10.0
sendgrid>=6.11.0
python-dotenv>=1.0.0
pandas>=2.0.0
